- Fixed ALL the loops using int instead of size_t - Fixed All warning in code. - Replace Thompson class with Graph, to better describe it. - Added State class to be used in Graph(UNFINISHED) - Added Transition class IN State, to be used in any state-like object. - Added comments for Graph, Transition and State. - I'm hungry.

+++ Project1/AST.cpp
@@ -23,7 +23,7 @@ BinTree AST::doASTree() {
 	BinTree *tempTree2 = new BinTree();
 
 	// since we know that the chain is valid because postFix is really OverPower...
-	for (int i = 0; i < _inputVector.size(); i++) {
+	for (size_t i = 0; i < _inputVector.size(); i++) {
 		// We need a switch statement to check every character in the input vector.
 		switch (_inputVector[i])
 		{
